refactor PUT /api/admin/restaurants/:id

+++ controllers/adminController.js
@@ -25,43 +25,14 @@ let adminController = {
       })
      },
     postRestaurant: (req, res) => {
-      if(!req.body.name){
-        req.flash('error_messages', "name didn't exist")
-        return res.redirect('back')
-      }
-    
-      const { file } = req
-      if (file) {
-        imgur.setClientID(IMGUR_CLIENT_ID);
-        imgur.upload(file.path, (err, img) => {
-          return Restaurant.create({
-            name: req.body.name,
-            tel: req.body.tel,
-            address: req.body.address,
-            opening_hours: req.body.opening_hours,
-            description: req.body.description,
-            image: file ? img.data.link : null,
-            CategoryId: req.body.categoryId
-          }).then((restaurant) => {
-            req.flash('success_messages', 'restaurant was successfully created')
-            return res.redirect('/admin/restaurants')
-          })
-        })
-      }
-      else {
-        return Restaurant.create({
-          name: req.body.name,
-          tel: req.body.tel,
-          address: req.body.address,
-          opening_hours: req.body.opening_hours,
-          description: req.body.description,
-          image: null,
-          CategoryId: req.body.categoryId
-        }).then((restaurant) => {
-          req.flash('success_messages', 'restaurant was successfully created')
-          return res.redirect('/admin/restaurants')
-        })
-       }
+      adminService.postRestaurant(req, res, (data) => {
+        if (data['status'] === 'error') {
+          req.flash('error_messages', data['message'])
+          return res.redirect('back')
+        }
+        req.flash('success_messages', data['message'])
+        res.redirect('/admin/restaurants')
+      })
     },
     getRestaurant: (req, res) => {
         adminService.getRestaurant(req, res, (data) => {
@@ -82,54 +53,15 @@ let adminController = {
       })
     },
 
-
-
     putRestaurant: (req, res) => {
-      if(!req.body.name){
-        req.flash('error_messages', "name didn't exist")
-        return res.redirect('back')
-      }
-    
-      const { file } = req
-      if (file) {
-        imgur.setClientID(IMGUR_CLIENT_ID);
-        imgur.upload(file.path, (err, img) => {
-          return Restaurant.findByPk(req.params.id)
-            .then((restaurant) => {
-              restaurant.update({
-                name: req.body.name,
-                tel: req.body.tel,
-                address: req.body.address,
-                opening_hours: req.body.opening_hours,
-                description: req.body.description,
-                image: file ? img.data.link : restaurant.image,
-                CategoryId: req.body.categoryId
-              })
-              .then((restaurant) => {
-                req.flash('success_messages', 'restaurant was successfully to update')
-                res.redirect('/admin/restaurants')
-              })
-            })
-        })
-      }
-      else {
-        return Restaurant.findByPk(req.params.id)
-          .then((restaurant) => {
-            restaurant.update({
-              name: req.body.name,
-              tel: req.body.tel,
-              address: req.body.address,
-              opening_hours: req.body.opening_hours,
-              description: req.body.description,
-              image: restaurant.image,
-              CategoryId: req.body.categoryId
-            })
-            .then((restaurant) => {
-              req.flash('success_messages', 'restaurant was successfully to update')
-              res.redirect('/admin/restaurants')
-            })
-          })
-      }
+    adminService.postRestaurant(req, res, (data) => {
+     if(data.status === 'error'){
+       req.flash('error_messages', data.message)
+       return res.redirect('back')
+     }
+     req.flash('success_messages', data.message)
+     res.redirect('/admin/restaurants')
+    })
     },
     deleteRestaurant: (req, res) => {
         adminService.deleteRestaurant(req, res, (data) => {
@@ -159,6 +91,7 @@ let adminController = {
       .catch(err => console.log(err))
   },
 }
+
 
 module.exports = adminController
